+++ core-ng-test/src/test/java/core/framework/test/TestModule.java
@@ -93,6 +93,7 @@ public class TestModule extends AbstractTestModule {
 
     private void configureKafka() {
         kafka().uri("kafka://localhost:9092");
+        kafka().maxMessageSize(950000);
         kafka().maxProcessTime(Duration.ofMinutes(30));
         kafka().longConsumerLagThreshold(Duration.ofSeconds(60));
         kafka().poolSize(1);