+++ core-ng/src/test/java/core/framework/internal/kafka/MessagePublisherImplTest.java
@@ -2,7 +2,6 @@ package core.framework.internal.kafka;
 
 import core.framework.internal.log.ActionLog;
 import core.framework.internal.log.LogManager;
-import core.framework.kafka.MessagePublisher;
 import core.framework.util.Strings;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -13,6 +12,7 @@ import java.util.List;
 
 import static java.nio.charset.StandardCharsets.UTF_8;
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.mockito.ArgumentMatchers.argThat;
 import static org.mockito.Mockito.verify;
 
@@ -20,13 +20,14 @@ import static org.mockito.Mockito.verify;
  * @author neo
  */
 class MessagePublisherImplTest {
-    private MessagePublisher<TestMessage> publisher;
+    private MessagePublisherImpl<TestMessage> publisher;
     private MessageProducer producer;
     private LogManager logManager;
 
     @BeforeEach
     void createMessagePublisher() {
         producer = Mockito.mock(MessageProducer.class);
+        producer.maxMessageSize = 1000;
         publisher = new MessagePublisherImpl<>(producer, "topic", TestMessage.class);
         logManager = new LogManager();
     }
@@ -51,5 +52,13 @@ class MessagePublisherImplTest {
         verify(producer).send(argThat(record -> Arrays.equals(Strings.bytes("key"), record.key()) && "topic".equals(record.topic())));
 
         logManager.end("end");
+    }
+
+    @Test
+    void validateMessageSize() {
+        producer.maxMessageSize = 9;
+        assertThatThrownBy(() -> publisher.validateMessageSize(Strings.bytes("12345"), Strings.bytes("12345")))
+                .isInstanceOf(Error.class)
+                .hasMessageContaining("message is too large");
     }
 }